LayoutTests/fast/dom/HTMLSelectElement/options-indexed-getter-precedence-expected.txt

Test that HTMLOptionsCollection's indexed properties take precedence over named properties

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS options.length is 2
PASS options[0].value is "option_1"
PASS options.item(0).value is "option_1"
PASS options[1].value is "option_2"
PASS options.item(1).value is "option_2"
PASS options[2] is undefined.
PASS options.item(2) is null
PASS options.namedItem('0').value is "option_2"
PASS successfullyParsed is true

TEST COMPLETE
